- Cleanup client cache code by implementing sync methods on top of async counterparts. - Refactored private internal implementation methods to AbstractClientInternalCacheProxy, removing intermediate class AbstractClientCacheProxy: now all Cache, ICache and ICacheInternal interface methods are implemented in ClientCacheProxy and invoke internal implementation in AbstractClientInternalCacheProxy. - Resolved a case in which when near cache was disabled, statistics would not be maintained even if enabled.
